feat: Add signOutFacebook function for logging out the user
This commit adds a new function, `signOutFacebook`, to the `app_function.dart` file. The function is responsible for signing out the user from both Firebase and Facebook. It first calls `FirebaseAuth.instance.signOut()` to sign out from Firebase, and then calls `FacebookAuth.instance.logOut()` to sign out from Facebook. After successful sign out, it prints a message indicating that the user has been logged out. In case of any error during sign out, it prints the error message.

This new function is useful for providing a way for users to log out of the app and ensures that they are signed out from both Firebase and Facebook platforms.