feat: Add Toastification package for displaying toast messages
This commit adds the `toastification` package to the project. The package provides a simple and customizable way to display toast messages in the app. Toast messages are used to show brief notifications or alerts to the user. By integrating this package, we can enhance the user experience by providing informative and visually appealing toast messages.

The `toastification` package is imported in the `main_page.dart` file, and a new import statement is added at the top of the file. This allows us to use the package's functionality throughout the app.

This feature is beneficial for providing important information or feedback to the user in a non-intrusive manner. It improves the overall usability and user engagement of the app.