feat: Update profile image upload process
This commit updates the profile image upload process in the `_ProfilePageState` class of `profile_page.dart`. It adds support for specifying the image type as 'profile' when calling the `updateImageUser` method in the `UserRepository`. Additionally, it modifies the `LoadProfile` event in the `UserBloc` to include the type as 'social' if the `isSocial` flag is present in the `novelBox`. These changes enhance the profile image management and loading functionality in the app.

Co-authored-by: [Author Name]