feat: Update notification icon in main_page.dart
This commit updates the notification icon in the `_MainpageState` class of `main_page.dart`. It replaces the `IconButton` widget with a `GestureDetector` widget to improve the user experience. The `onTap` callback is now used instead of the `onPressed` callback to handle the disconnection of the socket. Additionally, the `Icon` widget is wrapped in a `child` property to adhere to the Flutter widget structure. These changes enhance the functionality and usability of the notification feature in the app.