feat: Update dependencies for connectivity and nm
The commit updates the dependencies for `connectivity_plus` and `nm` in the `pubspec.lock` file. This ensures that the latest stable versions of these packages are used in the app. This enhancement improves the functionality and compatibility of the app with the latest features and bug fixes provided by these dependencies.