feat: Add hisRead property to ExpansionTileEpisode widget
This commit adds the `hisRead` property to the `ExpansionTileEpisode` widget in the `ExpansionTileEpisode.dart` file. The `hisRead` property is a list of dynamic type and is required for the widget. This enhancement improves the functionality of the `ExpansionTileEpisode` widget by providing the necessary data for rendering the user's reading history.