feat: Add isCheckPassword property to InputForm widget
This commit adds the `isCheckPassword` property to the `InputForm` widget in the `InputForm.dart` file. The `isCheckPassword` property is a boolean value that determines whether password validation should be applied. This enhancement improves the functionality of the `InputForm` widget by allowing developers to enable or disable password validation based on their requirements.