feat: Update user repository URL to use new API endpoint
The code changes in `user_repository.dart` update the `url` constant in the `UserRepository` class to use a new API endpoint. The previous endpoint was 'https://pzfbh88v-3002.asse.devtunnels.ms/api/', and it has been replaced with 'https://pzfbh88v-3004.asse.devtunnels.ms/api/'. This change ensures that the app communicates with the correct API server and improves the reliability and performance of the app.